requirements.txt
# Core dependencies for CISO-GENAI Framework
# Causal Credit Assignment for Multi-Agent Systems

# Numerical operations (required)
numpy>=1.20.0

# YAML configuration file parsing (required)
PyYAML>=6.0.0

# Optional: PyTorch for learnable policies
# Install separately if needed: pip install torch
# torch>=2.0.0

# Optional: Gymnasium for extended environment compatibility
# gymnasium>=0.29.0

